requirements.txt
cffi==1.6.0
Pillow==3.2.0
